some work on context assitance
